Allow gitit2 to work as a proper subsite.
This change allows gitit2 to work as a subsite with a Yesod scaffolded
application.

This change removes some routes from the subsite that would conflict
with routes that usually exist in a master site.

To do this in the sample application , I had to move the subsite out of
/ to a more specific directory, and then redirect / to gitit's subsite
default route.

There may be a better way of accomplishing this, but this seemed like
the obvious solution.